refactor: 게시물 컴포넌트 리팩토링
